ENH: add basic tests, fix debug index emit

+++ instrument/emit/Kernel.h
@@ -30,12 +30,12 @@ private:
     std::vector<DebugInfo> instrumentInstructions(const std::vector<llvm::Instruction*>& instructions);
     void emitDebugInfo(llvm::Function* function, const std::vector<DebugInfo>& debugRecods);
 
-    void instrumentStore(llvm::StoreInst* store, size_t debugIndex);
-    void instrumentLoad(llvm::LoadInst* load, size_t debugIndex);
+    void instrumentStore(llvm::StoreInst* store, int32_t debugIndex);
+    void instrumentLoad(llvm::LoadInst* load, int32_t debugIndex);
 
     bool isLocalStore(llvm::StoreInst* store);
     bool isLocalLoad(llvm::LoadInst* load);
 
     bool isInstrumentable(llvm::Instruction& instruction);
-    void instrumentInstruction(llvm::Instruction* instruction, size_t debugIndex);
+    void instrumentInstruction(llvm::Instruction* instruction, int32_t debugIndex);
 };
