+++ instrument/emit/MemoryAccess.h
@@ -17,8 +17,8 @@ public:
 
     }
 
-    void handleStore(llvm::StoreInst* store, size_t debugIndex);
-    void handleLoad(llvm::LoadInst* load, size_t debugIndex);
+    void handleStore(llvm::StoreInst* store, int32_t debugIndex);
+    void handleLoad(llvm::LoadInst* load, int32_t debugIndex);
 
 private:
     Context& context;
